getUserByUsername api

+++ src/main/java/com/photobucket/service/PostService.java
@@ -19,9 +19,16 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.photobucket.dao.CommentDao;
+import com.photobucket.dao.LikeDao;
 import com.photobucket.dao.PostDao;
 import com.photobucket.dao.UserDao;
 import com.photobucket.dto.PostDto;
+import com.photobucket.exception.NotFoundException;
+import com.photobucket.exception.PostNotFoundException;
+import com.photobucket.exception.UserNotFoundException;
+import com.photobucket.model.Comment;
+import com.photobucket.model.Like;
 import com.photobucket.model.Post;
 import com.photobucket.model.User;
 
@@ -35,12 +42,20 @@ public class PostService {
 	@Autowired
 	UserDao userDao;
 	
-	public ResponseEntity<?> deletePost(long id) throws PostServiceException {
+	@Autowired
+	CommentDao commentDao;
+	
+	@Autowired
+	LikeDao likeDao;
+	
+	public ResponseEntity<?> deletePost(long id){
 		Optional<Post> post = postRepo.findById(id);
+		
+		if(post.isEmpty()) {
+			throw new NotFoundException("Post Not Found !");
+		}
 		if(post.get().getId()>0) {
 			postRepo.delete(post.get());
-		}else {
-			throw new NoSuchElementException("Post Not Found !");
 		}
 		System.out.println("Post Deleted !");
 		return new ResponseEntity<String>("Post Deleted", HttpStatus.OK);
@@ -82,37 +97,49 @@ public class PostService {
     }
 	
 	public ResponseEntity<?> addPost(PostDto postDto) throws IOException {
-		Post post = new Post();
-		
-		if(postDto.getPostImg().isEmpty()) {
-			throw new java.lang.NullPointerException();
-		}
 		
 		long user_post_id = postDto.getUser_post_id();
 		Optional<User> user = userDao.findById(user_post_id);
-		post.setDescription(postDto.getDescription());
-		post.setTitle(postDto.getTitle());
-		post.setUser(user.get());
-		
-		Calendar calendar = Calendar.getInstance();
-		java.util.Date currentTime = calendar.getTime();
-		long time = currentTime.getTime();
-		
-		post.setDateTime(LocalDateTime.now());
-		byte[] compressedBytes = postDto.getPostImg().getBytes();
-		post.setImg(compressedBytes);
-		postRepo.save(post);
-		System.out.println("Post Saved !");
-		return new ResponseEntity<String>("Post Saved", HttpStatus.OK);
-		
+			if(user.isEmpty()) {
+				throw new NotFoundException("User not found !");
+			}
+			
+			if(user.get().isBlocked()) {
+				System.out.println("FROM IN"+user.get().isBlocked());
+				return new ResponseEntity<String>("Admin blocked your profile", HttpStatus.OK);
+			}else {
+				Post post = new Post();
+				
+				if(postDto.getPostImg().isEmpty()) {
+					throw new NotFoundException("Please Add Image ");
+				}
+				
+				System.out.println("Inside else");
+				post.setDescription(postDto.getDescription());
+				post.setTitle(postDto.getTitle());
+				post.setUser(user.get());
+				
+				Calendar calendar = Calendar.getInstance();
+				java.util.Date currentTime = calendar.getTime();
+				long time = currentTime.getTime();
+				
+				post.setDateTime(LocalDateTime.now());
+				byte[] compressedBytes = postDto.getPostImg().getBytes();
+				post.setImg(compressedBytes);
+				postRepo.save(post);
+				System.out.println("Post Saved !");
+				return new ResponseEntity<String>("Post Saved", HttpStatus.OK);
+			}
+
 	}
 	
 	public ResponseEntity<?> changePicture(long user_id, MultipartFile changeFile) throws IOException, PostServiceException {
 		Optional<Post> post = postRepo.findById(user_id);
-			
-		if(changeFile.isEmpty()) {
-			throw new java.lang.NullPointerException();
-		}
+			if(post.isEmpty()) {
+				throw new NotFoundException("Post Not Found !");
+			}else if(changeFile.isEmpty()) {
+				throw new NotFoundException("Please Add Image !");
+			}
 		
 			byte[] compressedBytes = changeFile.getBytes();
 			post.get().setImg(compressedBytes);
@@ -121,18 +148,66 @@ public class PostService {
 			return new ResponseEntity<String>("Picture Updated For Post",HttpStatus.OK);
 	}
 	
+	 public List<PostDto> getPostsByUser(long userId) {
+	        User user = userDao.findById(userId).get();
+	        List<Post> posts = user.getPosts();
+	        
+	        
+	        
+	        List<PostDto> postDTOs = new ArrayList<>();
+	        for (Post post : posts) {
+	        
+	        	PostDto postDTO = new PostDto();
+	            
+	            postDTO.setId(post.getId());
+	            postDTO.setTitle(post.getTitle());
+	            postDTO.setDescription(post.getDescription());
+//	            postDTO.setImg(post.getImg());
+	            postDTOs.add(postDTO);
+	        }
+	        return postDTOs;
+	    }
+	 
+	 public PostDto getPostImage(long postId) {
+	        Post post = postRepo.findById(postId).get();
+
+	        PostDto postDTO = new PostDto();
+	        postDTO.setId(post.getId());
+            postDTO.setImg(post.getImg());
+	        
+	        return postDTO;
+	    }
+	
+	
 	public ResponseEntity<?> getPost(long post_id) {
 		Optional<Post> post = postRepo.findById(post_id);
+		if(post.isEmpty()) {
+			throw new NotFoundException("Post not found !");
+		}
 			PostDto postDto = new PostDto();
 			postDto.setDescription(post.get().getDescription());
 			postDto.setTitle(post.get().getTitle());
 			postDto.setUser_post_id(post.get().getUser().getId());
 			postDto.setImg(post.get().getImg());
 			postDto.setId(post.get().getId());
-//			postDto.setComments(post.get().getComments());
+			
+			List<Comment> totalComment = commentDao.findByPostId(post_id);
+			
+			postDto.setComments(totalComment);
+			
+			List<Like> totalLike = likeDao.findByPostId(post_id);
+			postDto.setLikes(totalLike);
+			
 			System.out.println("Post Fetched");
 			return new ResponseEntity<PostDto>(postDto, HttpStatus.OK);
 	}
 	
+	public long getUser(String username) {
+		User user = userDao.findByUserName(username);
+		if(user.getId() <= 0) {
+			throw new NotFoundException("User Not Present");
+		}
+		return user.getId();
+	}
 	
 }
